Düzeltmeler yapıldı.

+++ src/main/java/com/findik/chatter/repository/IMessageRepository.java
@@ -1,6 +1,5 @@
 package com.findik.chatter.repository;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.findik.chatter.entity.Message;
 
 @Repository
-public interface IMessageRepository extends JpaRepository<Message, BigInteger> {
+public interface IMessageRepository extends JpaRepository<Message, Integer> {
 
 	List<Message> findAllByOrderByCreatedAtAsc();
 }
